Fuse the Movable and Grid classes.
